pyproject.toml
[project]
name = "erp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.21.0",
    "google-adk>=1.5.0",
    "litellm>=1.74.0.post1",
    "requests>=2.32.4",
]
